+++ src/app/App.tsx
@@ -4,16 +4,16 @@ import { MantineProvider } from '@mantine/core';
 import { Notifications } from '@mantine/notifications';
 import { useDisclosure } from '@mantine/hooks';
 
-import Layout from '../components/layout/Layout';
-import MainPage from '../components/pages/main/MainPage';
-import CatalogPage from '../components/pages/catalog/CatalogPage';
-import AboutPage from '../components/pages/about-us/AboutPage';
-import LoginPage from '../components/pages/login/LoginPage';
-import RegistrationPage from '../components/pages/registration/RegistrationPage';
-import BasketPage from '../components/pages/basket/BasketPage';
-import ProfilePage from '../components/pages/profile/ProfilePage';
-import NotFoundPage from '../components/pages/not-found/NotFoundPage';
-import ProtectedRoute from '../services/ProtectedRoute/ProtectedRoute';
+import Layout from '../layout/Layout';
+import MainPage from '../pages/main/MainPage';
+import CatalogPage from '../pages/catalog/CatalogPage';
+import AboutPage from '../pages/about-us/AboutPage';
+import LoginPage from '../pages/login/LoginPage';
+import RegistrationPage from '../pages/registration/RegistrationPage';
+import BasketPage from '../pages/basket/BasketPage';
+import ProfilePage from '../pages/profile/ProfilePage';
+import NotFoundPage from '../pages/not-found/NotFoundPage';
+import ProtectedRoute from '../routes/ProtectedRoute/ProtectedRoute';
 
 function App() {
   const loginState = localStorage.getItem('userLoggedIn');
