+++ src/layout/Layout.tsx
@@ -1,8 +1,8 @@
 import { Dispatch, SetStateAction } from 'react';
 import { Outlet } from 'react-router-dom';
 import { Container } from '@mantine/core';
-import { AppHeader } from './header/Header';
-import Footer from './footer/Footer';
+import { AppHeader } from '../components/header/Header';
+import Footer from '../components/footer/Footer';
 
 type LayoutProps = {
   setUserLoggedIn: Dispatch<SetStateAction<boolean>>;
