+++ src/styles/index.scss
@@ -1,4 +1,4 @@
 @import '../utils/libs/normalize.css';
 @import '../app/App.scss';
-@import '../components/layout/footer/Footer.scss';
+@import '../layout/footer/Footer.scss';
 @import './fonts.scss';
